requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1; python_version < '3.9'
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
django-allauth==0.51.0
dj_rest_auth
django==4.0.5
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pyjwt==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
